Move component file
