+++ src/pages/welcome.tsx
@@ -1,6 +1,6 @@
 import type { NextPage } from 'next'
 import Head from 'next/head'
-import NewSafe from '@/components/welcome'
+import NewSafe from '@/components/welcome/NewSafe'
 
 const Welcome: NextPage = () => {
   return (
